requirements.txt
bunch==1.0.1
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
enum34==1.1.6
Flask==0.12.2
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
Naked==0.1.31
pycrypto==2.6.1
PyYAML==3.12
redis==2.10.6
requests==2.18.4
shellescape==3.4.1
urllib3==1.22
Werkzeug==0.12.2
